refactor: move TargetResponse to targets.TargetResponse for better organization
