+++ docs/reference/target_response.md
@@ -1,1 +1,1 @@
-::: src.agenteval.target_response
+::: src.agenteval.targets.target_response
