+++ src/agenteval/__init__.py
@@ -10,9 +10,8 @@ from jinja2 import Environment, PackageLoader, select_autoescape
 from rich.logging import RichHandler
 
 from .hook import Hook
-from .target_response import TargetResponse
 
-__all__ = ["Hook", "TargetResponse"]
+__all__ = ["Hook"]
 __version__ = version("agent-evaluation")
