+++ src/agenteval/targets/__init__.py
@@ -1,8 +1,9 @@
 # Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 # SPDX-License-Identifier: Apache-2.0
 
+from .target_response import TargetResponse
 from .base_target import BaseTarget
 from .boto3_target import Boto3Target
 from .target_factory import TargetFactory
 
-__all__ = ["BaseTarget", "TargetFactory", "Boto3Target"]
+__all__ = ["TargetResponse", "BaseTarget", "TargetFactory", "Boto3Target"]
