+++ src/agenteval/targets/q_business/target.py
@@ -3,8 +3,7 @@
 
 from typing import Optional
 
-from agenteval import TargetResponse
-from agenteval.targets import Boto3Target
+from agenteval.targets import Boto3Target, TargetResponse
 
 _SERVICE_NAME = "qbusiness"
